* NROM checks

+++ include/vpnes/core/apu.hpp
@@ -162,9 +162,9 @@ public:
 	 * @param bus CPU bus
 	 */
 	void addHooksCPU(CBus &bus) {
-		for (std::size_t addr = 0x2000; addr < 0x4000; addr++) {
+		for (std::uint16_t addr = 0x2000; addr < 0x4000; addr++) {
 			bus.addPreReadHook(addr, *this, &CAPU::readReg);
-			bus.addWriteHook(addr, *this,  &CAPU::writeReg);
+			bus.addWriteHook(addr, *this, &CAPU::writeReg);
 		}
 	}
