+++ include/vpnes/core/ppu.hpp
@@ -234,7 +234,7 @@ public:
 	 * @param bus CPU bus
 	 */
 	void addHooksCPU(CBus &bus) {
-		for (std::size_t addr = 0x2000; addr < 0x4000; addr++) {
+		for (std::uint16_t addr = 0x2000; addr < 0x4000; addr++) {
 			bus.addPreReadHook(addr, *this, &CPPU::readReg);
 			bus.addWriteHook(addr, *this, &CPPU::writeReg);
 		}
